Refinement of the APP on its final stages of development
Commenting the rest of the code, adding some extra features related to aesthetics and to functionality

Developed a system of hiding the window and making ti appear again.

Tried out python programming for app development, it was deemed to be insuficient for my needs. So support for C++ development will be mantained